move page from addresses to account_activity module
