+++ src/account_activity/mod.rs
@@ -3,3 +3,4 @@ pub mod dialog;
 pub mod functions;
 mod graphql;
 pub mod models;
+pub mod page;
