+++ src/addresses/mod.rs
@@ -1,4 +1,2 @@
-pub mod functions;
 pub mod models;
-pub mod page;
 mod table_traits;
